+++ src/kernel/WeightedDegreeCharKernel.h
@@ -20,8 +20,8 @@
 class CWeightedDegreeCharKernel: public CSimpleKernel<CHAR>
 {
  public:
-  CWeightedDegreeCharKernel(LONG size, DREAL* weights, INT degree, INT max_mismatch, bool use_normalization=true, bool block_computation=false, INT mkl_stepsize=1) ;
-  CWeightedDegreeCharKernel(CCharFeatures* l, CCharFeatures* r, LONG size, DREAL* weights, INT degree, INT max_mismatch=0, bool use_normalization=true, bool block_computation=false, INT mkl_stepsize=1) ;
+  CWeightedDegreeCharKernel(INT size, DREAL* weights, INT degree, INT max_mismatch, bool use_normalization=true, bool block_computation=false, INT mkl_stepsize=1) ;
+  CWeightedDegreeCharKernel(CCharFeatures* l, CCharFeatures* r, INT size, DREAL* weights, INT degree, INT max_mismatch=0, bool use_normalization=true, bool block_computation=false, INT mkl_stepsize=1) ;
   ~CWeightedDegreeCharKernel() ;
   
   virtual bool init(CFeatures* l, CFeatures* r, bool do_init);
